catalog became a component + minor fixes
