+++ front/src/App.css
@@ -1,12 +1,3 @@
-.newcatalog .container {
-    display: grid;
-    grid-template-columns: 1fr 1fr 1fr;
-}
-
-.newcatalog {
-    padding-bottom: 5vw;
-}
-
 .App {
     background-image: url('./images/img.jpg');
 }
