+++ front/src/components/Catalog/Catalog.module.css
@@ -1,0 +1,5 @@
+.container {
+    display: grid;
+    grid-template-columns: 1fr 1fr 1fr;
+    padding-bottom: 5vh;
+}
