+++ front/src/pages/home.jsx
@@ -1,17 +1,17 @@
 import Slider from '../components/Slider/Slider';
-import new_films from '../staticStorage/new_films';
+
 import Films from '../components/Films/Films';
+import all_films from '../staticStorage/all_films';
+import c from '../components/Catalog/Catalog.module.css';
 
 function Home() {
     return (
         <div className="Home">
-            <Slider></Slider>
-            <div className="newcatalog">
-                <div className="container">
-                    {new_films.map((film) => (
-                        <Films key={film.id} work={film} />
-                    ))}
-                </div>
+            <Slider/>
+            <div className={c.container}>
+                {all_films.map((film) => (
+                    <Films key={film.id} work={film} />
+                ))}
             </div>
         </div>
     );
